added killed triggers
minor refactor on advancement formatting, and a failsafe for incomplete triggers occurring.

New triggers: "entity_killed_player" and "player_killed_entity". The ladder cannot run a function as the entity that was killed, to make up for the lack of functionality I have added a way to detect what type of entity was killed.

This is not on all triggers for performance reasons.